requirements.txt
aiohttp==3.8.3
aiohttp-sse-client==0.2.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
frozenlist==1.3.1
greenlet==2.0.0.post0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mnemonic==0.20
multidict==6.0.2
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==22.1.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.43
stellar-base-sseclient==0.0.21
stellar-sdk==8.1.1
toml==0.10.2
typeguard==2.13.3
urllib3==1.26.12
Werkzeug==2.2.2
yarl==1.8.1
